decoder: retry experiment where Bstring is a union type internally of String | ByteVector
This brings a few challenges, but the important thing to do this is to be ablet o encode the dictonary keys in a more correct way wherewe know we have already valid texttual keys coming from the bytestring.

This allows to remove the incorrect implmentation of Ordering that existed and do it in a more correct way only when we are sure the bytes actually where correct utf8 and so the comparisson only occurs when that is valid

Bute we gotta say working with union types like this sucks a lot, the DX is bad since we don't have something like rust with 'if let..' and the branches always need to be covered or verified, so it feels like going back in some ways.