feat: Enrich the constructors for the ast types
Ensure to provide ways to create valid ast values correctly by construction. Specially important for the bdictionary creation.

The other types are trivial, but the dicitionary has subtle problems and constraints that we want to ensure at construction time.

Some unsafe escape hatches are still provided to manipualte/create directly and invalid ast type, but those are useful for certain scenarious like tests or other things, like when we are sure of the values we are manipulating.

One specific case is the parsing, since we are doing our validation logic itself in the parser and don't allow invalid state coming from there, so by construction we know it's valid, so no need to use these safe methods that have more overhead.